test(form): add basic Form tests
