+++ package.json
@@ -32,7 +32,7 @@
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
     "@deploysentinel/cypress-debugger": "^0.2.5",
-    "@testing-library/cypress": "^7.0.2",
+    "@testing-library/cypress": "^8.0.2",
     "@types/fs-extra": "^8.0.1",
     "@types/jest": "^24.0.24",
     "@types/jquery": "^3.3.31",
@@ -221,7 +221,6 @@
     "@types/webpack": "^5.28.0",
     "react": "16.14.0",
     "react-dom": "16.14.0",
-    "pretty-format": "25.5.0",
     "typescript": "4.5.5",
     "codex-tooltip": "1.0.2",
     "@pulumi/pulumi": "3.25.0",
