+++ packages/form/package.json
@@ -16,7 +16,6 @@
   "dependencies": {
     "@babel/runtime": "^7.16.3",
     "@webiny/validation": "^5.27.0",
-    "invariant": "^2.2.4",
     "lodash": "^4.17.5"
   },
   "devDependencies": {
@@ -25,12 +24,15 @@
     "@babel/plugin-proposal-class-properties": "^7.16.0",
     "@babel/preset-env": "^7.16.4",
     "@babel/preset-typescript": "^7.16.0",
+    "@testing-library/react": "^12.1.5",
+    "@testing-library/user-event": "^12.8.3",
     "@types/invariant": "^2.2.35",
     "@types/lodash": "^4.14.178",
     "@webiny/cli": "^5.27.0",
     "@webiny/project-utils": "^5.27.0",
     "babel-plugin-lodash": "^3.3.4",
     "rimraf": "^3.0.2",
+    "ttypescript": "^1.5.13",
     "typescript": "4.5.5"
   },
   "peerDependencies": {
